@@ -1,5 +1,5 @@
 {
-  "name": "Officehours",
+  "name": "oh-client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -10,24 +10,17 @@
     "bootstrap": "^3.3.7",
     "jquery": "^3.3.1",
     "node-sass-chokidar": "^1.3.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-infinite-scroller": "^1.2.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
     "resolve": "^1.5.0",
     "socket.io-client": "^2.1.1",
     "toastr": "^2.1.4"
   },
-  "devDependencies": {
-    "del": "^3.0.0",
-    "react-infinite-scroller": "^1.1.3",
-    "request": "^2.81.0",
-    "request-promise": "^4.2.1"
-  },
   "scripts": {
-    "build-css": "node-sass-chokidar scss/ -o public/css/",
-    "watch-css": "npm run build-css && node-sass-chokidar scss/ -o public/css/ --watch --recursive",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
